Display how many stars can be sent + hide reminder if none
