+++ src/Thanks.php
@@ -73,17 +73,28 @@ class Thanks implements EventSubscriberInterface, PluginInterface
             return;
         }
 
+        $gitHub = new GitHubClient($event->getComposer(), $event->getIO());
+
+        $notStarred = 0;
+        foreach ($gitHub->getRepositories() as $repo) {
+            $notStarred += (int) !$repo['viewerHasStarred'];
+        }
+
+        if (!$notStarred) {
+            return;
+        }
+
         if ('Hyper' === getenv('TERM_PROGRAM')) {
             $love = '💖 ';
             $star = '⭐ ';
         } else {
-            $love = '\\' === DIRECTORY_SEPARATOR ? 'love' : '💖 ';
-            $star = '\\' === DIRECTORY_SEPARATOR ? 'star' : '★ ';            
+            $love = '\\' === \DIRECTORY_SEPARATOR ? 'love' : '💖 ';
+            $star = '\\' === \DIRECTORY_SEPARATOR ? 'star' : '★ ';
         }
 
         $this->io->writeError('');
         $this->io->writeError('What about running <comment>composer thanks</> now?');
-        $this->io->writeError(sprintf('This will spread some %s by sending a %s to the GitHub repositories of your fellow package maintainers.', $love, $star));
+        $this->io->writeError(sprintf('This will spread some %s by sending a %s to <comment>%d</comment> GitHub repositor%s of your fellow package maintainers.', $love, $star, $notStarred, 1 < $notStarred ? 'ies' : 'y'));
         $this->io->writeError('');
     }
